Pulls out logic for ability data diff to shared AbilityService
