+++ Danki2/Assets/Scripts/Abilities/Services/AbilityService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bb017d98465ecc4d91bc2db8e58dd0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
