+++ Danki2/Assets/Scripts/Actor/Actor.cs
@@ -32,10 +32,13 @@ public abstract class Actor : MonoBehaviour
         EffectManager = new EffectManager(this, updateSubject, statsManager);
         HealthManager = new HealthManager(this, updateSubject);
         InterruptionManager = new InterruptionManager();
-        ChannelService = new ChannelService(this, lateUpdateSubject, InterruptionManager);
-        InstantCastService = new InstantCastService(this);
-        InstantCastService.RegisterAbilityDataDiffer(new AbilityDataStatsDiffer(this));
         MovementManager = new MovementManager(this, updateSubject, navmeshAgent);
+
+        AbilityDataStatsDiffer abilityDataStatsDiffer = new AbilityDataStatsDiffer(this);
+        ChannelService = new ChannelService(this, lateUpdateSubject, InterruptionManager);
+        ChannelService.RegisterAbilityDataDiffer(abilityDataStatsDiffer);
+        InstantCastService = new InstantCastService(this);
+        InstantCastService.RegisterAbilityDataDiffer(abilityDataStatsDiffer);
 
         Dead = false;
     }
